Ensure bit integer values are returned as Int64
Fixes https://github.com/JuliaDatabases/SQLite.jl/issues/207. The issue
here was in our translation from sqlite type to Julia type, we were
using `Int`, though in the `sqlitevalue` function that actually
retrieved the value from the database, we used `sqlite3_column_int64`.
This change just ensures that even in 32-bit systems, we'll return
values as Int64, avoiding issues like the `RANDOM()` sqlite function,
which returns an Int64 value.

+++ src/SQLite.jl
@@ -366,7 +366,7 @@ end
 
 # convert SQLite stored type into Julia equivalent
 juliatype(x::Integer) =
-    x == SQLITE_INTEGER ? Int :
+    x == SQLITE_INTEGER ? Int64 :
     x == SQLITE_FLOAT ? Float64 :
     x == SQLITE_TEXT ? String :
     Any
@@ -377,7 +377,7 @@ function juliatype(decl_typestr::AbstractString,
                    default::Type = Any)
     typeuc = uppercase(decl_typestr)
     if typeuc in ("INTEGER", "INT")
-        return Int
+        return Int64
     elseif typeuc in ("NUMERIC", "REAL", "FLOAT")
         return Float64
     elseif typeuc == "TEXT"
